requirements.txt
opencv-python
python-dotenv
ultralytics
python-telegram-bot